Vessels passing through boundaries now works within the new model.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/Plugins/AngioFE2@9205 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ AngioFE2/Culture.cpp
@@ -94,7 +94,7 @@ Segment Culture::GrowSegment(Segment::TIP& tip, double starttime, double grow_ti
 	// determine the growth direction
 	vec3d seg_vec;
 	//now run it through the different filters
-	for (size_t i = 0; i < m_pmat->grow_direction_modifiers.size(); i++)
+	for (int i = 0; i < m_pmat->grow_direction_modifiers.size(); i++)
 	{
 		seg_vec = m_pmat->grow_direction_modifiers[i]->GrowModifyGrowDirection(seg_vec, tip, branch);
 	}
